make the library to support multi-resolutions without change the source
