+++ Adafruit_SSD1306_128_32.h
@@ -1,0 +1,40 @@
+/*********************************************************************
+This is a library for our Monochrome OLEDs based on SSD1306 drivers
+
+  Pick one up today in the adafruit shop!
+  ------> http://www.adafruit.com/category/63_98
+
+These displays use SPI to communicate, 4 or 5 pins are required to
+interface
+
+Adafruit invests time and resources providing this open source code,
+please support Adafruit and open-source hardware by purchasing
+products from Adafruit!
+
+Written by Limor Fried/Ladyada  for Adafruit Industries.
+BSD license, check license.txt for more information
+All text above, and the splash screen must be included in any redistribution
+*********************************************************************/
+#ifndef _Adafruit_SSD1306_128_32_H_
+#define _Adafruit_SSD1306_128_32_H_
+/*=========================================================================
+    SSD1306 Displays
+    -----------------------------------------------------------------------
+    The driver is used in multiple displays (128x64, 128x32, etc.).
+    Select the appropriate display below to create an appropriately
+    sized framebuffer, etc.
+
+    SSD1306_128_64  128x64 pixel display
+
+    SSD1306_128_32  128x32 pixel display
+
+    SSD1306_96_16
+
+    -----------------------------------------------------------------------*/
+//   #define SSD1306_128_64
+   #define SSD1306_128_32
+//   #define SSD1306_96_16
+/*=========================================================================*/
+#define Adafruit_SSD1306 Adafruit_SSD1306_128_32
+#include "Adafruit_SSD1306.core.h"
+#endif /* _Adafruit_SSD1306_H_ */
